+++ 2024-frontend/src/components/ObservatoryPurchases.vue
