+++ 2024-frontend/src/views/Observatoire.vue
@@ -6,7 +6,7 @@ import ObservatoryHero from "@/components/ObservatoryHero.vue"
 import ObservatoryFilters from "@/components/ObservatoryFilters.vue"
 import ObservatoryResultsFilters from "@/components/ObservatoryResultsFilters.vue"
 import ObservatoryResultsTop from "@/components/ObservatoryResultsTop.vue"
-import ObservatoryResultsPurchases from "@/components/ObservatoryResultsPurchases.vue"
+import ObservatoryPurchases from "@/components/ObservatoryPurchases.vue"
 
 /* Back to filters */
 const observatoryFilters = useTemplateRef("observatory-filters")
@@ -39,7 +39,7 @@ watchEffect(async () => {
       :teledeclarationsCount="stats.teledeclarationsCount"
       class="fr-mb-3w"
     />
-    <ObservatoryResultsPurchases />
+    <ObservatoryPurchases />
     <pre>{{ stats }}</pre>
   </section>
 </template>
